app/app/requirements.txt
Flask
mysql-connector-python
Flask-Cors